+++ apps/challenge-bot/src/lambda/handler.ts
@@ -2,7 +2,7 @@ import {
   createLambdaFunction,
   createProbot,
 } from '@probot/adapter-aws-lambda-serverless';
-import app from './index';
+import app from '../app';
 
 module.exports.webhooks = createLambdaFunction(app, {
   probot: createProbot(),
